Improve the guardband analysis script.
Remove the over_voltage_avs option as it is cannot be specified by the user. Instead, query the over_voltage_avs option and write the value to the CSV file. Make the output more verbose. Switch to the CPUFreq performance governor after the SoC has cooled down, respectively before the SoC is heated up.